Linked list updated with new functions.
Added list_copy makes deep copy of list.  Returns separate copy of the
not a new pointer to the old list.

Added list_print display contents of list to stdout. The list uses pointers
to void so the printout only includes the addresses of the data payload
and not the contents.

Function list_get_num returns pointer to node by number. Number zero
is the address of the head node.

Function list_node_swap reverses the data payload of two nodes.

Function list_reverse reverses the linked list payload. It should
be noted that the addresses of the nodes themselves are modified
during this operation.

Hash functions have been removed from unit tests. Hash functions not
complete in this revision and should not be used.